+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/actions/internal/ChannelActionsViewModel.kt
@@ -58,7 +58,7 @@ internal class ChannelActionsViewModel(
         chatClient.watchChannelAsState(
             cid = cid,
             messageLimit = DEFAULT_MESSAGE_LIMIT,
-            coroutineScope = viewModelScope
+            coroutineScope = viewModelScope,
         ).filterNotNull()
 
     /**