+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/viewmodel/factory/MessageListViewModelFactory.kt
@@ -28,7 +28,8 @@ import io.getstream.chat.android.ui.message.list.header.viewmodel.MessageListHea
  * A ViewModel factory for MessageListViewModel, MessageListHeaderViewModel and MessageInputViewModel.
  *
  * @param cid The channel id in the format messaging:123.
- * @param messageId The id of the target message to displayed.
+ * @param messageId The id of a message we wish to scroll to in messages list. Used to control the number of channel
+ * queries executed on screen initialization.
  *
  * @see MessageListViewModel
  * @see MessageListHeaderViewModel
@@ -40,10 +41,10 @@ public class MessageListViewModelFactory @JvmOverloads constructor(
 ) : ViewModelProvider.Factory {
 
     private val factories: Map<Class<*>, () -> ViewModel> = mapOf(
-        MessageListHeaderViewModel::class.java to { MessageListHeaderViewModel(cid) },
-        MessageInputViewModel::class.java to { MessageInputViewModel(cid) },
+        MessageListHeaderViewModel::class.java to { MessageListHeaderViewModel(cid, messageId = messageId) },
+        MessageInputViewModel::class.java to { MessageInputViewModel(cid, messageId = messageId) },
         MessageListViewModel::class.java to { MessageListViewModel(cid, messageId) },
-        MessageComposerViewModel::class.java to { MessageComposerViewModel(MessageComposerController(cid)) },
+        MessageComposerViewModel::class.java to { MessageComposerViewModel(MessageComposerController(cid, messageId = messageId)) },
     )
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {